+++ lib/phoenix/controller.ex
@@ -214,13 +214,15 @@ defmodule Phoenix.Controller do
         def call(conn, {:error, :not_found}) do
           conn
           |> put_status(:not_found)
-          |> render(MyErrorView, :"404")
+          |> put_view(MyErrorView)
+          |> render(:"404")
         end
 
         def call(conn, {:error, :unauthorized}) do
           conn
           |> put_status(403)
-          |> render(MyErrorView, :"403")
+          |> put_view(MyErrorView)
+          |> render(:"403")
         end
       end
   """
@@ -635,12 +637,8 @@ defmodule Phoenix.Controller do
 
   By default, Controllers render templates in a view with a similar name to the
   controller. For example, `MyApp.UserController` will render templates inside
-  the `MyApp.UserView`. This information can be changed any time by using
-  `render/3`, `render/4` or the `put_view/2` function:
-
-      def show(conn, _params) do
-        render(conn, MyApp.SpecialView, :show, message: "Hello")
-      end
+  the `MyApp.UserView`. This information can be changed any time by using the
+  `put_view/2` function:
 
       def show(conn, _params) do
         conn
@@ -704,10 +702,13 @@ defmodule Phoenix.Controller do
 
   def render(conn, view, template)
       when is_atom(view) and (is_binary(template) or is_atom(template)) do
+    IO.warn "#{__MODULE__}.render/3 with a view is deprecated, see the documentation for render/3 for an alternative"
     render(conn, view, template, [])
   end
 
   @doc """
+  WARNING: This function is deprecated in favor of `render/3` + `put_view/2`.
+
   A shortcut that renders the given template in the given view.
 
   Equivalent to:
@@ -720,6 +721,7 @@ defmodule Phoenix.Controller do
   @spec render(Plug.Conn.t, atom, atom | binary, Keyword.t | map) :: Plug.Conn.t
   def render(conn, view, template, assigns)
       when is_atom(view) and (is_binary(template) or is_atom(template)) do
+    IO.warn "#{__MODULE__}.render/4 with a view is deprecated, see the documentation for render/3 for an alternative"
     conn
     |> put_view(view)
     |> render(template, assigns)
@@ -780,12 +782,12 @@ defmodule Phoenix.Controller do
     |> send_resp(conn.status || default_status, body)
   end
 
-  defp ensure_resp_content_type(%{resp_headers: resp_headers} = conn, content_type) do
+  defp ensure_resp_content_type(%Plug.Conn{resp_headers: resp_headers} = conn, content_type) do
     if List.keyfind(resp_headers, "content-type", 0) do
       conn
     else
       content_type = content_type <> "; charset=utf-8"
-      %{conn | resp_headers: [{"content-type", content_type}|resp_headers]}
+      %Plug.Conn{conn | resp_headers: [{"content-type", content_type}|resp_headers]}
     end
   end
 
@@ -920,7 +922,7 @@ defmodule Phoenix.Controller do
     end
 
     params = Map.put(conn.params, required_key, param)
-    %{conn | params: params}
+    %Plug.Conn{conn | params: params}
   end
 
   defp scrub_param(%{__struct__: mod} = struct) when is_atom(mod) do