+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Enhancements
   * [Controller] Support partial file downloads with `:offset` and `:length` options to `send_download/3`
+  * [Logger] Add whitelist support to `filter_parameters` logger configuration, via new `:keep` tuple format
+  * [Router] Raise on duplicate plugs in pipe_through scopes
+
+* Deprecations
+  * [Controller] Passing a view in `render/3` and `render/4` is deprecated in favor of `put_view/2`
 
 ## 1.3.0 (2017-07-28)